+++ tests/views/fixtures/cross-field-validation-component.blade.php
@@ -1,0 +1,5 @@
+<div>
+    <input v-model="password" type="password">
+    <input v-model="password_confirmation" type="password">
+    <button v-click="submit">Submit</button>
+</div>